Hibernate Implementation completed
We improved and created several xml context files to use Hibernate ORM,
GenericEnumUserType and GenericHibernateDaoImpl where given by the
instructor so we didnt really touch those files, even so, the allow us
to create custom mapping in the orm file configuration

We Serialize OrdenCompra and thus we needed to change primitive type
into wrapped types. Only int was affect and so were the files that uses
them.

One little mistake is versioning the h2 database file, but it's not big
deal since the database does not grow up exponencially.